requirements.txt
# PDF table extraction
camelot-py[cv]==0.11.0

# Additional backends and utilities
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Export format support
openpyxl>=3.1.0  # For Excel export
pandas>=2.0.0    # Required by Camelot
